+++ libs/net/yolo/predict.py
@@ -4,8 +4,8 @@ import numpy as np
 import cv2
 import os
 import json
-from ...cython_utils.cy_yolo_findboxes import yolo_box_constructor
-from ...pascal_voc_io import PascalVocWriter, XML_EXT
+from ...cythonUtils.cy_yolo_findboxes import yolo_box_constructor
+from ...pascalVoc import PascalVocWriter, XML_EXT
 
 
 def _fix(obj, dims, scale, offs):
