+++ libs/ui/functions/helpFunctions.py
@@ -1,0 +1,6 @@
+import platform
+from libs.ui.functions.mainWindowFunctions import MainWindowFunctions
+
+
+class HelpFunctions(MainWindowFunctions):
+    pass
