+++ resources.qrc
@@ -36,6 +36,7 @@
 <file alias="demoWebcam">resources/icons/demoWebcam.png</file>
 <file alias="saveFile">resources/icons/save.png</file>
 <file alias="changeFormat">resources/icons/format_voc.png</file>
+<file alias="format_voc">resources/icons/format_voc.png</file>
 <file alias="format_yolo">resources/icons/format_yolo.png</file>
 <file alias="saveAs">resources/icons/save-as.png</file>
 <file alias="shapeFillColor">resources/icons/color.png</file>
@@ -52,5 +53,12 @@
 <file alias="strings">resources/strings/strings.properties</file>
 <file alias="strings-zh-TW">resources/strings/strings-zh-TW.properties</file>
 <file alias="strings-zh-CN">resources/strings/strings-zh-CN.properties</file>
+<file alias="strings-te-IN">resources/strings/strings-te-IN.properties</file>
+<file alias="strings-ru-RU">resources/strings/strings-ru-RU.properties</file>
+<file alias="strings-pt-BZ">resources/strings/strings-pt-BZ.properties</file>
+<file alias="strings-ja-JP">resources/strings/strings-ja-JP.properties</file>
+<file alias="strings-hi-IN">resources/strings/strings-hi-IN.properties</file>
+<file alias="strings-es-MX">resources/strings/strings-es-MX.properties</file>
+<file alias="strings-bn-IN">resources/strings/strings-bn-IN.properties</file>
 </qresource>
 </RCC>
