+++ resources/strings/strings-bn-IN.properties
@@ -1,0 +1,115 @@
+openFile=খোলা
+openFileDetail=চিত্র বা লেবেল ফাইল খুলুন
+close=অব্যাহতিপ্রাপ্ত
+closeDetail=প্রস্থান করুন
+openDir=ফোল্ডার খোলা
+openDirDetail=ফোল্ডার খোলা
+impVideo=ভিডিও ফ্রেম আমদানি করুন
+impVideoDetail=টিকা রচনার জন্য কাঁচা ভিডিও ফ্রেম আমদানি করুন
+commitAnnotatedFrames=প্রতিশ্রুতিবদ্ধ প্রতিশ্রুতিবদ্ধ
+commitAnnotatedFramesDetail=ডেটা ফোল্ডারে টীকাযুক্ত ফ্রেমগুলি ফেলে দিন
+trainModel=একটি মডেল প্রশিক্ষণ
+trainModelDetail=ডেটা ফোল্ডারে প্রশিক্ষণের জন্য একটি মডেল চয়ন করুন
+visualize=ভিজ্যুয়ালাইজ মডেল
+visualizeDetail=টেনসরবোর্ড ব্যবহার করে মডেলটি ভিজ্যুয়ালাইজ করুন
+frameByFrame=একটি ভিডিও এনেটেট করুন
+frameByFrameDetail=ফ্রেম ফ্রেমে টীকায়িত করতে একটি ভিডিও ফাইল চয়ন করুন
+demoWebcam=ওয়েবক্যামে ডেমো মডেল
+demoWebcamDetail=ওয়েবক্যাম ব্যবহার করে মডেলটি প্রদর্শন করুন
+changeSaveDirDetail=ডিফল্ট পরিবর্তন সংরক্ষিত টিকা ডায়র
+openAnnotation=ওপেন টিকা
+openAnnotationDetail=একটি টীকা ফাইল খুলুন
+changeSaveDir=ফোল্ডার সংরক্ষণ করুন
+nextImg=পরবর্তী চিত্র
+nextImgDetail=পরবর্তী চিত্র খুলুন
+prevImg=পূর্ববর্তী চিত্র
+prevImgDetail=পূর্ববর্তী চিত্রটি খুলুন
+verifyImg=চিত্র যাচাই করুন
+verifyImgDetail=চিত্র যাচাই করুন
+saveFile=সংরক্ষণ
+train=রেলগাড়ি
+predict=ভবিষ্যদ্বাণী করা
+annotate=টীকা লেখা
+analyze=বিশ্লেষণ করা
+selectProject=প্রকল্প নির্বাচন করুন
+projectLayout=প্রকল্পের নাম
+flowCmb=মোড
+modelCmb=মডেল
+loadCmb=চেক্পইণ্ট
+thresholdSpd=গোবরাট
+verbaliseChb=বাগাড়ম্বরপূর্ণ
+jsonChb=আউটপুট JSON টিকা
+vocChb=আউটপুট ভিওসি টিকা
+trainerCmb=প্রশিক্ষণ অ্যালগরিদম
+momentumSpd=ভরবেগ
+learningModeCmb=শেখার মোড
+learningRateSpd=প্রাথমিক শিক্ষার হার
+maxLearningRateSpd=সর্বোচ্চ শিক্ষার হার
+stepSizeCoefficient=পদক্ষেপের সহগ
+keepSpb=রাখার জন্য চেকপয়েন্টগুলি
+batchSpb=ব্যাচ আকার
+epochSpb=ইপোকস টু দ্য রান
+saveSpb=প্রত্যেকটি সংরক্ষণ করুন
+clipLayout=ক্লিপ গ্রেডিয়েন্টস
+updateAnchorChb=অ্যাঙ্কারস আপডেট করুন
+triangularTip=ডিফল্টরূপে, শিক্ষার হার প্রতিটি চক্রের ক্ষেত্রে রৈখিকভাবে বৃদ্ধি পায় এবং তারপরে রৈখিকভাবে হ্রাস পায়।
+triangular2Tip=শিক্ষার হারের পার্থক্য ব্যতীত ত্রিভুজ সংক্রান্ত নীতি হিসাবে একই প্রতিটি চক্রের শেষে অর্ধেক কাটা হয়। এর অর্থ প্রতিটি চক্রের পরে শিক্ষার হারের পার্থক্য হ্রাস পায়।
+exp_rangeTip=শিক্ষার হার ন্যূনতম এবং সর্বাধিক সীমাগুলির মধ্যে পরিবর্তিত হয় এবং প্রতিটি সীমানা মান হ্রাসের জন্য সূচকীয় উপাদান: ^ ^ গ্লোবাল_স্টেপ
+momentumTip=গতিবেগ এবং আরএমএসপ্রপ অপ্টিমাইজারের জন্য মুহুর্তের সেটিং
+saveFileDetail=লেবেলগুলিকে একটি ফাইলে সংরক্ষণ করুন
+changeFormatDetail=সংরক্ষণের ফর্ম্যাটটি পরিবর্তন করুন
+saveAs=সংরক্ষণ করুন
+saveAsDetail=ভিন্ন ফাইলে লেবেল সংরক্ষণ করুন
+closeFile=ঘনিষ্ঠ
+closeFileDetail=বর্তমান ফাইলটি বন্ধ করুন
+resetAll=সব পুনরায় সেট করুন
+resetAllDetail=সব পুনরায় সেট করুন
+boxLineColor=বক্স লাইন রঙ
+boxLineColorDetail=বক্স লাইনের রঙ চয়ন করুন
+createShape=বক্স তৈরি করুন
+createShapeDetail=একটি নতুন বাক্স আঁকুন
+setCreateMode=বক্স তৈরি করুন
+setCreateModeDetail=একটি নতুন বাক্স আঁকুন
+delBox=বক্স মুছুন
+delBoxDetail=বাক্সটি সরান
+copySelectedShape=সদৃশ বাক্স
+copySelectedShapeDetail=নির্বাচিত বাক্সের সদৃশ তৈরি করুন
+showTutorialDialog=অভিভাবকসংবঁধীয়
+showTutorialDialogDetail=ডেমো প্রদর্শন করুন
+showInfo=তথ্য
+showInfoDetail=অ্যাপ তথ্য
+zoomIn=প্রসারিত করো
+zoomInDetail=জুম স্তর বাড়ান
+zoomOut=ছোট করা
+zoomOutDetail=জুম স্তর হ্রাস করুন
+zoomOrg=মূল আকার
+zoomOrgDetail=মূল আকারে জুম করুন
+setFitWin=উইন্ডো ফিট করুন
+setFitWinDetail=জুম উইন্ডো আকার অনুসরণ করে
+setFitWidth=ফিট প্রস্থ
+setFitWidthDetail=জুম উইন্ডো প্রস্থ অনুসরণ করে
+editLabel=লেবেল সম্পাদনা করুন
+editLabelDetail=নির্বাচিত বাক্সের লেবেলটি সংশোধন করুন
+shapeLineColor=শেপ লাইন রঙ
+shapeLineColorDetail=এই নির্দিষ্ট আকারের জন্য লাইন রঙ পরিবর্তন করুন
+shapeFillColor=শেপ ফিল কালার
+shapeFillColorDetail=এই নির্দিষ্ট আকারের জন্য রঙ পূরণ করুন
+showHide=লেবেল প্যানেলটি দেখান / লুকান
+useDefaultLabel=ডিফল্ট লেবেল ব্যবহার করুন
+useDifficult=কঠিন
+boxLabelText=বক্স লেবেল
+labels=লেবেল
+autoSaveMode=অটো সেভ মোড
+singleClsMode=একক শ্রেণি মোড
+displayLabel=প্রদর্শন লেবেল
+fileList=ফাইল তালিকা
+files=নথি পত্র
+advancedMode=উন্নত মোড
+advancedModeDetail=উন্নত মোডে স্যুইচ করুন
+showAllDetail=সমস্ত বাউন্ডিং বাক্সগুলি দেখান
+showAll=বাউন্ডিং বক্স প্রদর্শন করুন
+hideAllDetail=সমস্ত আবদ্ধ বাক্স লুকান
+hideAll=বাউন্ডিং বক্স লুকান
+changeFormat=& পাস্কালভোক
+setEditMode=সম্পাদনা বাক্স
+setEditModeDetail=বাক্সগুলি সরান এবং সম্পাদনা করুন
