+++ resources/strings/strings-es-MX.properties
@@ -1,0 +1,115 @@
+openFile=Abierto
+openFileDetail=Abrir archivo de imagen o etiqueta
+close=Dejar
+closeDetail=Salir de la aplicación
+openDir=Carpeta abierta
+openDirDetail=Carpeta abierta
+impVideo=Importar cuadros de video
+impVideoDetail=Importar fotogramas de video sin formato para anotación
+commitAnnotatedFrames=Cometer anotaciones
+commitAnnotatedFramesDetail=Volcar marcos anotados en la carpeta de datos
+trainModel=Entrenar a un modelo
+trainModelDetail=Elija un modelo para entrenar en la carpeta de datos
+visualize=Visualizar modelo
+visualizeDetail=Visualice el modelo usando el tensorboard
+frameByFrame=Anotar un video
+frameByFrameDetail=Elija un archivo de video para anotar cuadro por cuadro
+demoWebcam=Modelo de demostración en la webcam
+demoWebcamDetail=Demuestre el modelo usando una cámara web
+changeSaveDirDetail=Cambiar el directorio de Anotación guardado predeterminado
+openAnnotation=Anotación abierta
+openAnnotationDetail=Abrir un archivo de anotaciones
+changeSaveDir=Cambiar Guardar Carpeta
+nextImg=Imagen siguiente
+nextImgDetail=Abre la siguiente imagen
+prevImg=Imagen anterior
+prevImgDetail=Abrir la imagen anterior
+verifyImg=Verificar imagen
+verifyImgDetail=Verificar imagen
+saveFile=Salvar
+train=Entrenar
+predict=Predecir
+annotate=Anotar
+analyze=Analizar
+selectProject=Seleccionar proyecto
+projectLayout=Nombre del proyecto
+flowCmb=Modo
+modelCmb=Modelo
+loadCmb=Control
+thresholdSpd=Límite
+verbaliseChb=Verboso
+jsonChb=Salida de anotaciones JSON
+vocChb=Salida de anotaciones de VOC
+trainerCmb=Algoritmo de entrenamiento
+momentumSpd=Impulso
+learningModeCmb=Modo de aprendizaje
+learningRateSpd=Tasa de aprendizaje inicial
+maxLearningRateSpd=Tasa máxima de aprendizaje
+stepSizeCoefficient=Coeficiente de tamaño de paso
+keepSpb=Puntos de control para mantener
+batchSpb=Tamaño del lote
+epochSpb=Épocas para correr
+saveSpb=Guardar cada
+clipLayout=Gradientes de clip
+updateAnchorChb=Actualizar anclas
+triangularTip=Por defecto, la tasa de aprendizaje aumenta linealmente en cada ciclo y luego disminuye linealmente.
+triangular2Tip=Lo mismo que la política triangular, excepto que la diferencia en la tasa de aprendizaje se reduce a la mitad al final de cada ciclo. Esto significa que la diferencia en la tasa de aprendizaje cae después de cada ciclo.
+exp_rangeTip=La tasa de aprendizaje varía entre los límites mínimo y máximo, y el factor exponencial para cada caída de valor límite es: γ ^ global_step
+momentumTip=Configuración de momento para impulso y optimizador rmsprop
+saveFileDetail=Guarda las etiquetas en un archivo
+changeFormatDetail=Cambiar formato de guardado
+saveAs=Guardar como
+saveAsDetail=Guarda las etiquetas en un archivo diferente
+closeFile=Cerca
+closeFileDetail=Cerrar el archivo actual
+resetAll=Resetear todo
+resetAllDetail=Resetear todo
+boxLineColor=Color de línea de caja
+boxLineColorDetail=Elegir color de línea de caja
+createShape=Crear caja
+createShapeDetail=Dibuja una nueva caja
+setCreateMode=Crear caja
+setCreateModeDetail=Dibuja una nueva caja
+delBox=Eliminar cuadro
+delBoxDetail=Quitar la caja
+copySelectedShape=Caja duplicada
+copySelectedShapeDetail=Crea un duplicado del cuadro seleccionado
+showTutorialDialog=Tutorial
+showTutorialDialogDetail=Mostrar demo
+showInfo=Información
+showInfoDetail=Información de la aplicación
+zoomIn=Acercarse
+zoomInDetail=Aumenta el nivel de zoom
+zoomOut=Disminuir el zoom
+zoomOutDetail=Disminuir el nivel de zoom
+zoomOrg=Tamaño original
+zoomOrgDetail=Zoom al tamaño original
+setFitWin=Ajustar ventana
+setFitWinDetail=El zoom sigue el tamaño de la ventana
+setFitWidth=Ajuste ancho
+setFitWidthDetail=El zoom sigue el ancho de la ventana
+editLabel=Editar etiqueta
+editLabelDetail=Modificar la etiqueta del cuadro seleccionado
+shapeLineColor=Color de línea de forma
+shapeLineColorDetail=Cambiar el color de línea para esta forma específica
+shapeFillColor=Color de relleno de forma
+shapeFillColorDetail=Cambiar el color de relleno para esta forma específica
+showHide=Mostrar / Ocultar panel de etiquetas
+useDefaultLabel=Usar etiqueta predeterminada
+useDifficult=difícil
+boxLabelText=Etiquetas de caja
+labels=Etiquetas
+autoSaveMode=Modo de guardado automático
+singleClsMode=Modo de clase individual
+displayLabel=Mostrar etiquetas
+fileList=Lista de archivos
+files=Archivos
+advancedMode=Modo avanzado
+advancedModeDetail=Cambiar a modo avanzado
+showAllDetail=Mostrar todos los cuadros delimitadores
+showAll=Mostrar cuadro delimitador
+hideAllDetail=Ocultar todos los cuadros delimitadores
+hideAll=Ocultar cuadro delimitador
+changeFormat=& PascalVOC
+setEditMode=Cuadro de edición
+setEditModeDetail=Mover y editar cuadros
