+++ resources/strings/strings-hi-IN.properties
@@ -1,0 +1,115 @@
+openFile=खुला हुआ
+openFileDetail=ओपन इमेज या लेबल फाइल
+close=छोड़ना
+closeDetail=आवेदन छोड़ो
+openDir=फोल्डर खोलो
+openDirDetail=फोल्डर खोलो
+impVideo=वीडियो फ्रेम आयात करें
+impVideoDetail=एनोटेशन के लिए कच्चे वीडियो फ्रेम आयात करें
+commitAnnotatedFrames=कमिटमेंट एनोटेशन
+commitAnnotatedFramesDetail=डंप ने डेटा फ़ोल्डर में फ़्रेम को एनोटेट किया
+trainModel=एक मॉडल को प्रशिक्षित करें
+trainModelDetail=डेटा फ़ोल्डर पर प्रशिक्षित करने के लिए एक मॉडल चुनें
+visualize=मॉडल की कल्पना करें
+visualizeDetail=टेंसरबोर्ड का उपयोग करके मॉडल की कल्पना करें
+frameByFrame=एक वीडियो एनोटेट करें
+frameByFrameDetail=फ्रेम द्वारा एनोटेट करने के लिए एक वीडियो फ़ाइल चुनें
+demoWebcam=वेबकैम पर डेमो मॉडल
+demoWebcamDetail=वेबकैम का उपयोग करके मॉडल प्रदर्शित करें
+changeSaveDirDetail=डिफ़ॉल्ट रूप से सहेजे गए एनोटेशन डायर को बदलें
+openAnnotation=ओपन एनोटेशन
+openAnnotationDetail=एनोटेशन फ़ाइल खोलें
+changeSaveDir=सहेजें फ़ोल्डर बदलें
+nextImg=अगली छवि
+nextImgDetail=अगली छवि खोलें
+prevImg=प्रीव इमेज
+prevImgDetail=पिछला चित्र खोलें
+verifyImg=छवि सत्यापित करें
+verifyImgDetail=छवि सत्यापित करें
+saveFile=सहेजें
+train=रेल गाडी
+predict=भविष्यवाणी
+annotate=एन्नोटेट
+analyze=विश्लेषण
+selectProject=प्रोजेक्ट का चयन करें
+projectLayout=परियोजना का नाम
+flowCmb=मोड
+modelCmb=नमूना
+loadCmb=जांच की चौकी
+thresholdSpd=द्वार
+verbaliseChb=वाचाल
+jsonChb=आउटपुट JSON एनोटेशन
+vocChb=आउटपुट VOC एनोटेशन
+trainerCmb=प्रशिक्षण एल्गोरिथ्म
+momentumSpd=गति
+learningModeCmb=लर्निंग मोड
+learningRateSpd=प्रारंभिक सीखने की दर
+maxLearningRateSpd=अधिकतम सीखने की दर
+stepSizeCoefficient=चरण आकार गुणांक
+keepSpb=रखने के लिए चौकी
+batchSpb=बैच का आकार
+epochSpb=युगों तक चला
+saveSpb=हर बचाओ
+clipLayout=क्लिप ग्रेजुएट्स
+updateAnchorChb=एंकर अपडेट करें
+triangularTip=डिफ़ॉल्ट रूप से, सीखने की दर प्रत्येक चक्र में रैखिक रूप से बढ़ जाती है और फिर रैखिक रूप से घट जाती है।
+triangular2Tip=प्रत्येक चक्र के अंत में सीखने की दर के अंतर को छोड़कर त्रिकोणीय नीति को आधा में काटा जाता है। इसका अर्थ है कि प्रत्येक चक्र के बाद सीखने की दर का अंतर कम हो जाता है।
+exp_rangeTip=सीखने की दर न्यूनतम और अधिकतम सीमाओं के बीच भिन्न होती है, और प्रत्येक सीमा मूल्य ड्रॉप के लिए घातांक कारक है: ep ^ __epep
+momentumTip=गति और rmsprop अनुकूलक के लिए गति सेटिंग
+saveFileDetail=फ़ाइल में लेबल सहेजें
+changeFormatDetail=सहेजें प्रारूप बदलें
+saveAs=के रूप रक्षित करें
+saveAsDetail=लेबल को किसी भिन्न फ़ाइल में सहेजें
+closeFile=बंद करे
+closeFileDetail=वर्तमान फ़ाइल बंद करें
+resetAll=सभी को रीसेट करें
+resetAllDetail=सभी को रीसेट करें
+boxLineColor=बॉक्स लाइन का रंग
+boxLineColorDetail=बॉक्स लाइन रंग चुनें
+createShape=बॉक्स बनाएँ
+createShapeDetail=एक नया बॉक्स ड्रा करें
+setCreateMode=बॉक्स बनाएँ
+setCreateModeDetail=एक नया बॉक्स ड्रा करें
+delBox=हटाएँ बॉक्स
+delBoxDetail=बॉक्स को हटा दें
+copySelectedShape=डुप्लिकेट बॉक्स
+copySelectedShapeDetail=चयनित बॉक्स का एक डुप्लिकेट बनाएं
+showTutorialDialog=ट्यूटोरियल
+showTutorialDialogDetail=प्रदर्शन दिखाएँ
+showInfo=जानकारी
+showInfoDetail=ऐप की जानकारी
+zoomIn=ज़ूम इन
+zoomInDetail=ज़ूम स्तर बढ़ाएँ
+zoomOut=ज़ूम आउट
+zoomOutDetail=ज़ूम स्तर घटाएँ
+zoomOrg=मूल आकार
+zoomOrgDetail=मूल आकार में ज़ूम करें
+setFitWin=फिट खिड़की
+setFitWinDetail=ज़ूम विंडो के आकार का अनुसरण करता है
+setFitWidth=चौड़ाई पर फ़िट
+setFitWidthDetail=ज़ूम विंडो की चौड़ाई का अनुसरण करता है
+editLabel=लेबल संपादित करें
+editLabelDetail=चयनित बॉक्स के लेबल को संशोधित करें
+shapeLineColor=आकार रेखा रंग
+shapeLineColorDetail=इस विशिष्ट आकार के लिए लाइन रंग बदलें
+shapeFillColor=आकार भरें रंग
+shapeFillColorDetail=इस विशिष्ट आकार के लिए रंग भरें
+showHide=लेबल पैनल दिखाएँ / छिपाएँ
+useDefaultLabel=डिफ़ॉल्ट लेबल का उपयोग करें
+useDifficult=कठिन
+boxLabelText=बॉक्स लेबल
+labels=लेबल
+autoSaveMode=ऑटो सेव मोड
+singleClsMode=सिंगल क्लास मोड
+displayLabel=लेबल प्रदर्शित करें
+fileList=फ़ाइल सूची
+files=फ़ाइलें
+advancedMode=उन्नत स्थिति
+advancedModeDetail=उन्नत मोड में जाएं
+showAllDetail=सभी बाउंडिंग बॉक्स दिखाएं
+showAll=बाउंडिंग बॉक्स दिखाएं
+hideAllDetail=सभी बाउंडिंग बॉक्स छिपाएं
+hideAll=बाउंडिंग बॉक्स छिपाएँ
+changeFormat=और पास्कलवॉक
+setEditMode=बॉक्स संपादित करें
+setEditModeDetail=बक्से को स्थानांतरित करें और संपादित करें
