+++ resources/strings/strings-ja-JP.properties
@@ -1,0 +1,115 @@
+openFile=開いた
+openFileDetail=画像またはラベルファイルを開く
+close=終了する
+closeDetail=アプリケーションを終了
+openDir=開いたフォルダ
+openDirDetail=開いたフォルダ
+impVideo=ビデオフレームのインポート
+impVideoDetail=注釈用の未加工のビデオフレームをインポートする
+commitAnnotatedFrames=注釈をコミット
+commitAnnotatedFramesDetail=注釈付きフレームをデータフォルダーにダンプする
+trainModel=モデルをトレーニングする
+trainModelDetail=データフォルダーでトレーニングするモデルを選択します
+visualize=モデルを視覚化
+visualizeDetail=テンソルボードを使用してモデルを可視化する
+frameByFrame=ビデオに注釈を付ける
+frameByFrameDetail=フレームごとに注釈を付けるビデオファイルを選択してください
+demoWebcam=ウェブカメラのデモモデル
+demoWebcamDetail=ウェブカメラを使用してモデルをデモする
+changeSaveDirDetail=保存されているデフォルトの注釈ディレクトリを変更
+openAnnotation=注釈を開く
+openAnnotationDetail=注釈ファイルを開く
+changeSaveDir=保存フォルダを変更
+nextImg=次の画像
+nextImgDetail=次の画像を開く
+prevImg=前の画像
+prevImgDetail=前の画像を開く
+verifyImg=画像を確認
+verifyImgDetail=画像を確認
+saveFile=保存する
+train=列車
+predict=予測する
+annotate=注釈を付ける
+analyze=分析する
+selectProject=プロジェクトを選択
+projectLayout=プロジェクト名
+flowCmb=モード
+modelCmb=型番
+loadCmb=チェックポイント
+thresholdSpd=敷居
+verbaliseChb=冗長
+jsonChb=JSONアノテーションの出力
+vocChb=出力VOCアノテーション
+trainerCmb=訓練アルゴリズム
+momentumSpd=勢い
+learningModeCmb=学習モード
+learningRateSpd=初期学習率
+maxLearningRateSpd=最大学習率
+stepSizeCoefficient=ステップサイズ係数
+keepSpb=保持するチェックポイント
+batchSpb=バッチサイズ
+epochSpb=実行するエポック
+saveSpb=すべてを保存
+clipLayout=グラデーションをクリップ
+updateAnchorChb=アンカーを更新
+triangularTip=デフォルトでは、学習率は各サイクルで直線的に増加し、その後直線的に減少します。
+triangular2Tip=学習率の差が各サイクルの終わりに半分にされることを除いて、三角政策と同じです。これは、各サイクルの後で学習率の差が減少することを意味します。
+exp_rangeTip=学習率は最小境界と最大境界の間で変化し、各境界値のドロップの指数係数はγ^global_stepです。
+momentumTip=運動量およびrmspropオプティマイザの運動量設定
+saveFileDetail=ラベルをファイルに保存する
+changeFormatDetail=保存形式を変更
+saveAs=名前を付けて保存
+saveAsDetail=ラベルを別のファイルに保存する
+closeFile=閉じる
+closeFileDetail=現在のファイルを閉じる
+resetAll=すべてリセット
+resetAllDetail=すべてリセット
+boxLineColor=ボックスラインカラー
+boxLineColorDetail=ボックスラインの色を選択してください
+createShape=ボックスを作成
+createShapeDetail=新しいボックスを描く
+setCreateMode=ボックスを作成
+setCreateModeDetail=新しいボックスを描く
+delBox=ボックスを削除
+delBoxDetail=箱を外す
+copySelectedShape=複製ボックス
+copySelectedShapeDetail=選択したボックスの複製を作成します
+showTutorialDialog=チュートリアル
+showTutorialDialogDetail=デモを表示
+showInfo=情報
+showInfoDetail=アプリ情報
+zoomIn=ズームイン
+zoomInDetail=ズームレベルを上げる
+zoomOut=ズームアウトする
+zoomOutDetail=ズームレベルを下げる
+zoomOrg=オリジナルサイズ
+zoomOrgDetail=元のサイズにズーム
+setFitWin=ウィンドウに合わせる
+setFitWinDetail=ズームはウィンドウサイズに従います
+setFitWidth=幅に合わせる
+setFitWidthDetail=ウィンドウの幅に合わせてズーム
+editLabel=ラベルを編集
+editLabelDetail=選択したボックスのラベルを変更します
+shapeLineColor=シェイプラインカラー
+shapeLineColorDetail=この特定の形状の線の色を変更します
+shapeFillColor=シェイプフィルの色
+shapeFillColorDetail=この特定の形状の塗りつぶし色を変更します
+showHide=ラベルパネルの表示/非表示
+useDefaultLabel=デフォルトのラベルを使用
+useDifficult=難しい
+boxLabelText=ボックスラベル
+labels=ラベル
+autoSaveMode=自動保存モード
+singleClsMode=シングルクラスモード
+displayLabel=ラベルを表示
+fileList=ファイルリスト
+files=ファイル
+advancedMode=アドバンストモード
+advancedModeDetail=詳細モードに切り替え
+showAllDetail=すべての境界ボックスを表示
+showAll=境界ボックスを表示
+hideAllDetail=すべての境界ボックスを非表示にする
+hideAll=境界ボックスを非表示
+changeFormat=＆PascalVOC
+setEditMode=ボックスを編集
+setEditModeDetail=ボックスの移動と編集
