+++ resources/strings/strings-pt-BZ.properties
@@ -1,0 +1,115 @@
+openFile=Abrir
+openFileDetail=Abrir imagem ou arquivo de etiqueta
+close=Sair
+closeDetail=Sair da aplicação
+openDir=Pasta aberta
+openDirDetail=Pasta aberta
+impVideo=Importar quadros de vídeo
+impVideoDetail=Importar quadros de vídeo brutos para anotação
+commitAnnotatedFrames=Anotar anotações
+commitAnnotatedFramesDetail=Despejar quadros anotados na pasta de dados
+trainModel=Treinar um modelo
+trainModelDetail=Escolha um modelo para treinar na pasta de dados
+visualize=Visualize Model
+visualizeDetail=Visualize o modelo usando o tensorboard
+frameByFrame=Anotar um vídeo
+frameByFrameDetail=Escolha um arquivo de vídeo para anotar quadro a quadro
+demoWebcam=Modelo de demonstração na webcam
+demoWebcamDetail=Demonstre o modelo usando uma webcam
+changeSaveDirDetail=Alterar dir padrão de anotação salvo
+openAnnotation=Anotação aberta
+openAnnotationDetail=Abrir um arquivo de anotação
+changeSaveDir=Alterar Salvar Pasta
+nextImg=Imagem seguinte
+nextImgDetail=Abra a próxima imagem
+prevImg=Imagem anterior
+prevImgDetail=Abra a imagem anterior
+verifyImg=Verificar imagem
+verifyImgDetail=Verificar imagem
+saveFile=Salve 
+train=Trem
+predict=Prever
+annotate=Anotar
+analyze=Analisar
+selectProject=Selecionar Projeto
+projectLayout=Nome do Projeto
+flowCmb=Modo
+modelCmb=Modelo
+loadCmb=Ponto de Verificação
+thresholdSpd=Limite
+verbaliseChb=Verbose
+jsonChb=Anotações JSON de saída
+vocChb=Output VOC Annotations
+trainerCmb=Algoritmo de Treinamento
+momentumSpd=Momentum
+learningModeCmb=Modo de Aprendizagem
+learningRateSpd=Taxa de Aprendizagem Inicial
+maxLearningRateSpd=Taxa Máxima de Aprendizagem
+stepSizeCoefficient=Coeficiente de tamanho da etapa
+keepSpb=Pontos de verificação a manter
+batchSpb=Tamanho do batch
+epochSpb=Épocas a serem executadas
+saveSpb=Salvar todos
+clipLayout=Gradientes de clipes
+updateAnchorChb=Atualizar âncoras
+triangularTip=Por padrão, a taxa de aprendizado aumenta linearmente em cada ciclo e depois diminui linearmente.
+triangular2Tip=O mesmo que a política triangular, exceto a diferença da taxa de aprendizado, é cortada pela metade no final de cada ciclo. Isso significa que a diferença na taxa de aprendizado cai após cada ciclo.
+exp_rangeTip=A taxa de aprendizado varia entre os limites mínimo e máximo, e o fator exponencial para cada queda de valor do limite é: γ^global_step
+momentumTip=Configuração de momento para otimizador de momento e rmsprop
+saveFileDetail=Salve os rótulos em um arquivo
+changeFormatDetail=Alterar o formato de salvamento
+saveAs=Salvar como
+saveAsDetail=Salve os rótulos em um arquivo diferente
+closeFile=Perto
+closeFileDetail=Feche o arquivo atual
+resetAll=Reiniciar tudo
+resetAllDetail=Reiniciar tudo
+boxLineColor=Cor da linha da caixa
+boxLineColorDetail=Escolha a cor da linha da caixa
+createShape=Criar caixa
+createShapeDetail=Desenhe uma nova caixa
+setCreateMode=Criar caixa
+setCreateModeDetail=Desenhe uma nova caixa
+delBox=Excluir caixa
+delBoxDetail=Retire a caixa
+copySelectedShape=Caixa Duplicada
+copySelectedShapeDetail=Crie uma duplicata da caixa selecionada
+showTutorialDialog=Tutorial
+showTutorialDialogDetail=Mostrar demonstração
+showInfo=Em formação
+showInfoDetail=Informações do aplicativo
+zoomIn=Mais Zoom
+zoomInDetail=Aumentar o nível de zoom
+zoomOut=Reduzir o zoom
+zoomOutDetail=Diminuir o nível de zoom
+zoomOrg=Tamanho original
+zoomOrgDetail=Aplicar zoom ao tamanho original
+setFitWin=Janela de ajuste
+setFitWinDetail=O zoom segue o tamanho da janela
+setFitWidth=Largura de ajuste
+setFitWidthDetail=O zoom segue a largura da janela
+editLabel=Editar etiqueta
+editLabelDetail=Modifique o rótulo da caixa selecionada
+shapeLineColor=Forma Linha Cor
+shapeLineColorDetail=Alterar a cor da linha para esta forma específica
+shapeFillColor=Forma Cor de preenchimento
+shapeFillColorDetail=Alterar a cor de preenchimento para esta forma específica
+showHide=Mostrar / ocultar painel de etiquetas
+useDefaultLabel=Usar etiqueta padrão
+useDifficult=difícil
+boxLabelText=Etiquetas da caixa
+labels=Etiquetas
+autoSaveMode=Modo de salvamento automático
+singleClsMode=Modo Classe Única
+displayLabel=Etiquetas de exibição
+fileList=Lista de Arquivos
+files=arquivos
+advancedMode=Advanced Mode
+advancedModeDetail=Mudar para o modo avançado
+showAllDetail=Mostrar todas as caixas delimitadoras
+showAll=Mostrar caixa delimitadora
+hideAllDetail=Ocultar todas as caixas delimitadoras
+hideAll=Ocultar caixa delimitadora
+changeFormat=& PascalVOC
+setEditMode=Caixa de edição
+setEditModeDetail=Mover e editar caixas
