+++ resources/strings/strings-ru-RU.properties
@@ -1,0 +1,115 @@
+openFile=открыто
+openFileDetail=Открыть файл изображения или метки
+close=Уволиться
+closeDetail=Выйти из приложения
+openDir=Открыть папку
+openDirDetail=Открыть папку
+impVideo=Импорт видеокадров
+impVideoDetail=Импорт необработанных видеокадров для аннотации
+commitAnnotatedFrames=Фиксация аннотаций
+commitAnnotatedFramesDetail=Дамп аннотированных кадров в папку данных
+trainModel=Тренировать модель
+trainModelDetail=Выберите модель для тренировки на папке данных
+visualize=Визуализировать модель
+visualizeDetail=Визуализация модели с использованием тензорной доски
+frameByFrame=Аннотировать видео
+frameByFrameDetail=Выберите видеофайл, чтобы комментировать кадр за кадром
+demoWebcam=Демо-модель на веб-камеру
+demoWebcamDetail=Продемонстрировать модель с помощью веб-камеры
+changeSaveDirDetail=Изменить сохраненный по умолчанию каталог аннотаций
+openAnnotation=Открытая аннотация
+openAnnotationDetail=Откройте файл аннотации
+changeSaveDir=Изменить Сохранить папку
+nextImg=Следующее изображение
+nextImgDetail=Откройте следующее изображение
+prevImg=Предыдущее изображение
+prevImgDetail=Откройте предыдущее изображение
+verifyImg=Проверить изображение
+verifyImgDetail=Проверить изображение
+saveFile=Сохранить
+train=Поезд
+predict=прогнозировать
+annotate=Пометки
+analyze=анализировать
+selectProject=Выберите проект
+projectLayout=название проекта
+flowCmb=Режим
+modelCmb=Модель
+loadCmb=Пропускной пункт
+thresholdSpd=порог
+verbaliseChb=Подробный
+jsonChb=Выходные JSON Аннотации
+vocChb=Выходные аннотации VOC
+trainerCmb=Алгоритм обучения
+momentumSpd=инерция
+learningModeCmb=Режим обучения
+learningRateSpd=Начальная скорость обучения
+maxLearningRateSpd=Максимальная скорость обучения
+stepSizeCoefficient=Коэффициент размера шага
+keepSpb=Контрольные точки, чтобы сохранить
+batchSpb=Размер партии
+epochSpb=Эпохи бега
+saveSpb=Сохранить каждый
+clipLayout=Клип Градиенты
+updateAnchorChb=Обновление якорей
+triangularTip=По умолчанию скорость обучения увеличивается линейно в каждом цикле, а затем уменьшается линейно.
+triangular2Tip=Аналогично треугольной политике, за исключением того, что разница в скорости обучения сокращается пополам в конце каждого цикла. Это означает, что разница в скорости обучения падает после каждого цикла.
+exp_rangeTip=Скорость обучения варьируется между минимальной и максимальной границами, и экспоненциальный коэффициент для каждого падения граничного значения: γ ^ global_step
+momentumTip=Настройка импульса для оптимизатора импульса и rmsprop
+saveFileDetail=Сохранить ярлыки в файл
+changeFormatDetail=Изменить формат сохранения
+saveAs=Сохранить как
+saveAsDetail=Сохраните метки в другой файл
+closeFile=близко
+closeFileDetail=Закрыть текущий файл
+resetAll=Сбросить все
+resetAllDetail=Сбросить все
+boxLineColor=Цвет линии коробки
+boxLineColorDetail=Выберите цвет линии Box
+createShape=Создать коробку
+createShapeDetail=Нарисуйте новую коробку
+setCreateMode=Создать коробку
+setCreateModeDetail=Нарисуйте новую коробку
+delBox=Удалить поле
+delBoxDetail=Снимите коробку
+copySelectedShape=Дубликат Box
+copySelectedShapeDetail=Создать копию выбранного поля
+showTutorialDialog=Руководство
+showTutorialDialogDetail=Показать демо
+showInfo=Информация
+showInfoDetail=Информация о приложении
+zoomIn=Увеличить
+zoomInDetail=Увеличить уровень масштабирования
+zoomOut=Уменьшить
+zoomOutDetail=Уменьшить уровень масштабирования
+zoomOrg=Первоначальный размер
+zoomOrgDetail=Увеличить до оригинального размера
+setFitWin=Fit Window
+setFitWinDetail=Увеличение следует за размером окна
+setFitWidth=Подходит ширина
+setFitWidthDetail=Увеличение следует за шириной окна
+editLabel=Изменить ярлык
+editLabelDetail=Изменить метку выбранного ящика
+shapeLineColor=Цвет линии формы
+shapeLineColorDetail=Изменить цвет линии для этой конкретной формы
+shapeFillColor=Цвет заливки формы
+shapeFillColorDetail=Изменить цвет заливки для этой конкретной формы
+showHide=Показать / Скрыть панель меток
+useDefaultLabel=Использовать метку по умолчанию
+useDifficult=трудно
+boxLabelText=Ярлыки коробки
+labels=Этикетки
+autoSaveMode=Режим автосохранения
+singleClsMode=Режим одного класса
+displayLabel=Ярлыки дисплея
+fileList=Список файлов
+files=файлы
+advancedMode=Расширенный режим
+advancedModeDetail=Переключиться в расширенный режим
+showAllDetail=Показать все ограничивающие рамки
+showAll=Показать ограничивающую рамку
+hideAllDetail=Скрыть все ограничивающие рамки
+hideAll=Скрыть ограничивающую рамку
+changeFormat=& PascalVOC
+setEditMode=Редактировать окно
+setEditModeDetail=Переместить и редактировать поля
