+++ resources/strings/strings-te-IN.properties
@@ -1,0 +1,115 @@
+openFile=ఓపెన్
+openFileDetail=చిత్రం లేదా లేబుల్ ఫైల్‌ను తెరవండి
+close=క్విట్
+closeDetail=అప్లికేషన్ నుండి నిష్క్రమించండి
+openDir=ఫోల్డర్ను తెరువు
+openDirDetail=ఫోల్డర్ను తెరువు
+impVideo=వీడియో ఫ్రేమ్‌లను దిగుమతి చేయండి
+impVideoDetail=ఉల్లేఖన కోసం ముడి వీడియో ఫ్రేమ్‌లను దిగుమతి చేయండి
+commitAnnotatedFrames=ఉల్లేఖనాలను కమిట్ చేయండి
+commitAnnotatedFramesDetail=ఉల్లేఖన ఫ్రేమ్‌లను డేటా ఫోల్డర్‌లో వేయండి
+trainModel=మోడల్‌కు శిక్షణ ఇవ్వండి
+trainModelDetail=డేటా ఫోల్డర్‌లో శిక్షణ ఇవ్వడానికి మోడల్‌ను ఎంచుకోండి
+visualize=మోడల్‌ను విజువలైజ్ చేయండి
+visualizeDetail=టెన్సార్‌బోర్డ్ ఉపయోగించి మోడల్‌ను విజువలైజ్ చేయండి
+frameByFrame=వీడియోను ఉల్లేఖించండి
+frameByFrameDetail=ఫ్రేమ్ ద్వారా ఫ్రేమ్‌ను ఉల్లేఖించడానికి వీడియో ఫైల్‌ను ఎంచుకోండి
+demoWebcam=వెబ్‌క్యామ్‌లో డెమో మోడల్
+demoWebcamDetail=వెబ్‌క్యామ్ ఉపయోగించి మోడల్‌ను ప్రదర్శించండి
+changeSaveDirDetail=డిఫాల్ట్ సేవ్ చేసిన ఉల్లేఖనాన్ని మార్చండి dir
+openAnnotation=ఉల్లేఖనాన్ని తెరవండి
+openAnnotationDetail=ఉల్లేఖన ఫైల్‌ను తెరవండి
+changeSaveDir=సేవ్ ఫోల్డర్ మార్చండి
+nextImg=తదుపరి చిత్రం
+nextImgDetail=తదుపరి చిత్రాన్ని తెరవండి
+prevImg=మునుపటి చిత్రం
+prevImgDetail=మునుపటి చిత్రాన్ని తెరవండి
+verifyImg=చిత్రాన్ని ధృవీకరించండి
+verifyImgDetail=చిత్రాన్ని ధృవీకరించండి
+saveFile=సేవ్
+train=రైలు
+predict=ఊహించండి
+annotate=వ్యాఖ్యానించడం
+analyze=విశ్లేషించడానికి
+selectProject=ప్రాజెక్ట్ ఎంచుకోండి
+projectLayout=ప్రాజెక్ట్ పేరు
+flowCmb=మోడ్
+modelCmb=మోడల్
+loadCmb=తనిఖీ స్థానం
+thresholdSpd=త్రెష్
+verbaliseChb=మందమైన
+jsonChb=అవుట్పుట్ JSON ఉల్లేఖనాలు
+vocChb=అవుట్పుట్ VOC ఉల్లేఖనాలు
+trainerCmb=శిక్షణ అల్గోరిథం
+momentumSpd=ఊపందుకుంటున్నది
+learningModeCmb=లెర్నింగ్ మోడ్
+learningRateSpd=ప్రారంభ అభ్యాస రేటు
+maxLearningRateSpd=గరిష్ట అభ్యాస రేటు
+stepSizeCoefficient=దశ పరిమాణం గుణకం
+keepSpb=ఉంచడానికి చెక్‌పాయింట్లు
+batchSpb=గుంపు పరిమాణం
+epochSpb=అమలు చేయడానికి యుగాలు
+saveSpb=ప్రతి సేవ్
+clipLayout=క్లిప్ ప్రవణతలు
+updateAnchorChb=యాంకర్లను నవీకరించండి
+triangularTip=అప్రమేయంగా, అభ్యాస రేటు ప్రతి చక్రంలో సరళంగా పెరుగుతుంది మరియు తరువాత సరళంగా తగ్గుతుంది.
+triangular2Tip=ప్రతి చక్రం చివరిలో అభ్యాస రేటు వ్యత్యాసం మినహా త్రిభుజాకార విధానం సగానికి తగ్గించబడుతుంది. ప్రతి చక్రం తరువాత అభ్యాస రేటు వ్యత్యాసం పడిపోతుందని దీని అర్థం.
+exp_rangeTip=అభ్యాస రేటు కనీస మరియు గరిష్ట సరిహద్దుల మధ్య మారుతూ ఉంటుంది మరియు ప్రతి సరిహద్దు విలువ తగ్గింపుకు ఘాతాంక కారకం: ^ ^ global_step
+momentumTip=మొమెంటం మరియు rmsprop ఆప్టిమైజర్ కోసం మొమెంటం సెట్టింగ్
+saveFileDetail=లేబుల్‌లను ఫైల్‌కు సేవ్ చేయండి
+changeFormatDetail=సేవ్ ఆకృతిని మార్చండి
+saveAs=ఇలా సేవ్ చేయండి
+saveAsDetail=లేబుల్‌లను వేరే ఫైల్‌కు సేవ్ చేయండి
+closeFile=దగ్గరగా
+closeFileDetail=ప్రస్తుత ఫైల్‌ను మూసివేయండి
+resetAll=అన్నీ రీసెట్ చేయండి
+resetAllDetail=అన్నీ రీసెట్ చేయండి
+boxLineColor=బాక్స్ లైన్ రంగు
+boxLineColorDetail=బాక్స్ లైన్ రంగును ఎంచుకోండి
+createShape=పెట్టెను సృష్టించండి
+createShapeDetail=క్రొత్త పెట్టెను గీయండి
+setCreateMode=పెట్టెను సృష్టించండి
+setCreateModeDetail=క్రొత్త పెట్టెను గీయండి
+delBox=బాక్స్ తొలగించు
+delBoxDetail=పెట్టెను తొలగించండి
+copySelectedShape=నకిలీ పెట్టె
+copySelectedShapeDetail=ఎంచుకున్న పెట్టె యొక్క నకిలీని సృష్టించండి
+showTutorialDialog=ట్యుటోరియల్
+showTutorialDialogDetail=డెమో చూపించు
+showInfo=సమాచారం
+showInfoDetail=అనువర్తన సమాచారం
+zoomIn=పెద్దదిగా చూపు
+zoomInDetail=జూమ్ స్థాయిని పెంచండి
+zoomOut=పెద్దది చెయ్యి
+zoomOutDetail=జూమ్ స్థాయిని తగ్గించండి
+zoomOrg=అసలు పరిమాణం
+zoomOrgDetail=అసలు పరిమాణానికి జూమ్ చేయండి
+setFitWin=విండోను అమర్చండి
+setFitWinDetail=జూమ్ విండో పరిమాణాన్ని అనుసరిస్తుంది
+setFitWidth=ఫిట్ వెడల్పు
+setFitWidthDetail=జూమ్ విండో వెడల్పును అనుసరిస్తుంది
+editLabel=లేబుల్‌ను సవరించండి
+editLabelDetail=ఎంచుకున్న పెట్టె యొక్క లేబుల్‌ను సవరించండి
+shapeLineColor=షేప్ లైన్ కలర్
+shapeLineColorDetail=ఈ నిర్దిష్ట ఆకారం కోసం పంక్తి రంగును మార్చండి
+shapeFillColor=షేప్ ఫిల్ కలర్
+shapeFillColorDetail=ఈ నిర్దిష్ట ఆకారం కోసం పూరక రంగును మార్చండి
+showHide=లేబుల్ ప్యానెల్ చూపించు / దాచు
+useDefaultLabel=డిఫాల్ట్ లేబుల్ ఉపయోగించండి
+useDifficult=కష్టం
+boxLabelText=బాక్స్ లేబుల్స్
+labels=Labels
+autoSaveMode=ఆటో సేవ్ మోడ్
+singleClsMode=సింగిల్ క్లాస్ మోడ్
+displayLabel=డిస్ప్లే లేబుల్స్
+fileList=ఫైల్ జాబితా
+files=ఫైళ్లు
+advancedMode=ఆధునిక పద్ధతి
+advancedModeDetail=అధునాతన మోడ్‌కు మారండి
+showAllDetail=అన్ని సరిహద్దు పెట్టెలను చూపించు
+showAll=బౌండింగ్ బాక్స్ చూపించు
+hideAllDetail=అన్ని సరిహద్దు పెట్టెలను దాచండి
+hideAll=సరిహద్దు పెట్టెను దాచు
+changeFormat=& పాస్కల్వోసి
+setEditMode=పెట్టెను సవరించండి
+setEditModeDetail=పెట్టెలను తరలించండి మరియు సవరించండి
