+++ resources/strings/strings-zh-CN.properties
@@ -1,46 +1,43 @@
-saveAsDetail=將标签保存到其他文件
-changeSaveDir=改变存放目录
-openFile=打开文件
+openFile=打开
+openFileDetail=打开图像或标签文件
+close=退出
+closeDetail=退出申请
+openDir=打开文件夹
+openDirDetail=打开文件夹
 impVideo=导入视频帧
-impVideoDetail=导入视频帧以进行评论
-shapeLineColorDetail=更改线条颜色
-resetAll=全部重置
-createShape=创建区块
-createShapeDetail=创建一个新的区块
-setCreateMode=创建区块
-setCreateModeDetail=创建一个新的区块
-copySelectedShapeDetail=复制区块
-verifyImg=验证图像
-zoomInDetail=放大
-verifyImgDetail=验证图像
-saveFileDetail=保存标签文件
-openFileDetail=打开图像文件
-setFitWidthDetail=调整宽度适应到窗口宽度
-showTutorialDialog=YouTube教学
-editLabel=编辑标签
-openAnnotationDetail=打开标签文件
-quit=退出
-shapeFillColorDetail=更改填充颜色
-closeFileDetail=关闭当前文件
-closeFile=关闭文件
-setFitWin=调整到窗口大小
-delBox=删除选择的区块
-boxLineColorDetail=选择线框颜色
-zoomOrg=原始大小
-resetAllDetail=重置所有设定
-zoomOutDetail=放大画面
+impVideoDetail=导入原始视频帧进行注释
+commitAnnotatedFrames=提交注释
+commitAnnotatedFramesDetail=将带注释的帧转储到数据文件夹中
+trainModel=训练模型
+trainModelDetail=选择要在数据文件夹上训练的模型
+visualize=可视化模型
+visualizeDetail=使用张量板可视化模型
+frameByFrame=注释视频
+frameByFrameDetail=选择一个视频文件以逐帧注释
+demoWebcam=网络摄像头上的演示模型
+demoWebcamDetail=使用网络摄像头演示模型
+changeSaveDirDetail=更改默认保存的注释目录
+openAnnotation=公开注解
+openAnnotationDetail=打开注释文件
+changeSaveDir=更改保存文件夹
+nextImg=下一张图片
+nextImgDetail=打开下一张图片
+prevImg=上一张图片
+prevImgDetail=打开上一张图片
+verifyImg=验证图片
+verifyImgDetail=验证图片
 saveFile=保存
-train=训练
+train=培养
 predict=预测
-annotate=诠
-analyze=剖析
-selectProject=选择你的项目
+annotate=注释
+analyze=分析
+selectProject=选择项目
 projectLayout=项目名称
 flowCmb=模式
 modelCmb=模型
 loadCmb=检查站
-thresholdSpd=门槛值
-verbaliseChb=罗唆
+thresholdSpd=阈
+verbaliseChb=详细
 jsonChb=输出JSON注释
 vocChb=输出VOC注释
 trainerCmb=训练算法
@@ -56,50 +53,63 @@ saveSpb=保存每一个
 clipLayout=剪辑渐变
 updateAnchorChb=更新锚点
 triangularTip=默认情况下，学习率在每个循环中线性增加，然后线性降低。
-triangular2Tip=与三角政策相同，但学习率差异除外在每个周期的末尾减少一半。这意味着学习率差异在每个周期后都会降低。
-exp_rangeTip=学习率在最小和最大边界之间变化，每个边界值下降的指数因子为：γ^全局步
+triangular2Tip=除了学习率差异在每个周期结束时减半外，其他与三角策略相同。这意味着学习率差异在每个周期后都会下降。
+exp_rangeTip=学习率在最小和最大边界之间变化，每个边界值下降的指数因子为：γ^ global_step
 momentumTip=动量和rmsprop优化器的动量设置
+saveFileDetail=将标签保存到文件
+changeFormatDetail=更改保存格式
 saveAs=另存为
-setFitWinDetail=缩放到当前窗口大小
-openDir=打开目录
-openDirDetail=打开目录
-showHide=显示/隐藏标签
-changeFormatDetail=更改存储格式
-shapeFillColor=填充颜色
-closeDetail=退出程序
-copySelectedShape=复制区块
-delBoxDetail=删除区块
-zoomIn=放大画面
+saveAsDetail=将标签保存到其他文件
+closeFile=关
+closeFileDetail=关闭当前文件
+resetAll=重置所有
+resetAllDetail=重置所有
+boxLineColor=箱线颜色
+boxLineColorDetail=选择框线颜色
+createShape=创建盒子
+createShapeDetail=画一个新盒子
+setCreateMode=创建盒子
+setCreateModeDetail=画一个新盒子
+delBox=删除框
+delBoxDetail=取下盒子
+copySelectedShape=重复盒
+copySelectedShapeDetail=创建所选框的副本
+showTutorialDialog=讲解
+showTutorialDialogDetail=显示演示
 showInfo=信息
-showInfoDetail=信息
-openAnnotation=开启标签
-prevImgDetail=上一个图像
-setFitWidth=缩放到跟当前画面一样宽
-zoomOut=缩小画面
-changeSaveDirDetail=更改保存标签文件的预设目录
-nextImgDetail=下一个图像
-zoomOrgDetail=放大到原始大小
-prevImg=上一个图像
-showTutorialDialogDetail=显示示范内容
-shapeLineColor=形状线条颜色
-boxLineColor=区块线条颜色
-editLabelDetail=修改当前所选的区块颜色
-nextImg=下一个图片
-useDefaultLabel=使用预设标签
-useDifficult=有难度的
-boxLabelText=区块的标签
+showInfoDetail=应用程式资讯
+zoomIn=放大
+zoomInDetail=增加缩放等级
+zoomOut=缩小
+zoomOutDetail=降低缩放等级
+zoomOrg=原始尺寸
+zoomOrgDetail=缩放到原始大小
+setFitWin=适合窗口
+setFitWinDetail=缩放跟随窗口大小
+setFitWidth=适合宽度
+setFitWidthDetail=缩放跟随窗口宽度
+editLabel=编辑标签
+editLabelDetail=修改所选盒子的标签
+shapeLineColor=形状线颜色
+shapeLineColorDetail=更改此特定形状的线条颜色
+shapeFillColor=形状填充颜色
+shapeFillColorDetail=更改此特定形状的填充颜色
+showHide=显示/隐藏标签面板
+useDefaultLabel=使用默认标签
+useDifficult=难
+boxLabelText=盒子标签
 labels=标签
 autoSaveMode=自动保存模式
-singleClsMode=单一类别模式
-displayLabel=显示类别
-fileList=文件列表
-files=文件
-advancedMode=专家模式
-advancedModeDetail=切换到专家模式
-showAllBoxDetail=显示所有区块
-hideAllBoxDetail=隐藏所有区块
-changeFormat=&PascalVOC
-setEditMode=
-setEditModeDetail=
-commitAnnotatedFrames=提交标签
-commitAnnotatedFramesDetail=将带注释的图像提交到数据文件夹
+singleClsMode=单班模式
+displayLabel=显示标签
+fileList=档案清单
+files=档案
+advancedMode=高级模式
+advancedModeDetail=切换至进阶模式
+showAllDetail=显示所有边界框
+showAll=显示边界框
+hideAllDetail=隐藏所有边界框
+hideAll=隐藏边界框
+changeFormat=＆PascalVOC
+setEditMode=编辑框
+setEditModeDetail=移动和编辑框
