+++ resources/strings/strings-zh-TW.properties
@@ -1,46 +1,43 @@
-saveAsDetail=將標籤保存到其他文件
-changeSaveDir=改變存放目錄
-openFile=開啟檔案
+openFile=打開
+openFileDetail=打開圖像或標籤文件
+close=退出
+closeDetail=退出申請
+openDir=打開文件夾
+openDirDetail=打開文件夾
 impVideo=導入視頻幀
-impVideoDetail=導入視頻幀以供評論
-shapeLineColorDetail=更改線條顏色
-resetAll=重置
-createShape=創建區塊
-createShapeDetail=畫一個區塊
-setCreateMode=創建區塊
-setCreateModeDetail=畫一個區塊
-copySelectedShapeDetail=複製區塊
-verifyImg=驗證圖像
-zoomInDetail=放大
-verifyImgDetail=驗證圖像
-saveFileDetail=將標籤存到
-openFileDetail=打開圖像
-setFitWidthDetail=調整到窗口寬度
-showTutorialDialog=YouTube教學
-editLabel=編輯標籤
-openAnnotationDetail=打開標籤文件
-quit=結束
-shapeFillColorDetail=更改填充顏色
-closeFileDetail=關閉目前檔案
-closeFile=關閉
-setFitWin=調整到跟窗口一樣大小
-delBox=刪除選取區塊
-boxLineColorDetail=選擇框線顏色
-zoomOrg=原始大小
-resetAllDetail=重設所有設定
-zoomOutDetail=畫面放大
-saveFile=儲存
-train=訓練
+impVideoDetail=導入原始視頻幀進行註釋
+commitAnnotatedFrames=提交註釋
+commitAnnotatedFramesDetail=將帶註釋的幀轉儲到數據文件夾中
+trainModel=訓練模型
+trainModelDetail=選擇要在數據文件夾上訓練的模型
+visualize=可視化模型
+visualizeDetail=使用張量板可視化模型
+frameByFrame=註釋視頻
+frameByFrameDetail=選擇一個視頻文件以逐幀註釋
+demoWebcam=網絡攝像頭上的演示模型
+demoWebcamDetail=使用網絡攝像頭演示模型
+changeSaveDirDetail=更改默認保存的註釋目錄
+openAnnotation=公開註解
+openAnnotationDetail=打開註釋文件
+changeSaveDir=更改保存文件夾
+nextImg=下一張圖片
+nextImgDetail=打開下一張圖片
+prevImg=上一張圖片
+prevImgDetail=打開上一張圖片
+verifyImg=驗證圖片
+verifyImgDetail=驗證圖片
+saveFile=保存
+train=培養
 predict=預測
 annotate=註釋
 analyze=分析
-selectProject=選擇你的項目
+selectProject=選擇項目
 projectLayout=項目名稱
 flowCmb=模式
 modelCmb=模型
 loadCmb=檢查站
-thresholdSpd=門檻值
-verbaliseChb=囉唆
+thresholdSpd=閾
+verbaliseChb=詳細
 jsonChb=輸出JSON註釋
 vocChb=輸出VOC註釋
 trainerCmb=訓練算法
@@ -56,50 +53,63 @@ saveSpb=保存每一個
 clipLayout=剪輯漸變
 updateAnchorChb=更新錨點
 triangularTip=默認情況下，學習率在每個循環中線性增加，然後線性降低。
-triangular2Tip=與三角策略相同，不同之處在於在每個週期結束時學習率的差異減少了一半。這意味著學習率差異將在每個週期後減小。
-exp_rangeTip=學習率在最小和最大邊界之間變化，每個邊界值下降的指數因子為：γ^全局步
+triangular2Tip=除了學習率差異在每個週期結束時減半外，其他與三角策略相同。這意味著學習率差異在每個週期後都會下降。
+exp_rangeTip=學習率在最小和最大邊界之間變化，每個邊界值下降的指數因子為：γ^global_step
 momentumTip=動量和rmsprop優化器的動量設置
+saveFileDetail=將標籤保存到文件
+changeFormatDetail=更改保存格式
 saveAs=另存為
-setFitWinDetail=縮放到窗口一樣
-openDir=開啟目錄
-openDirDetail=開啟目錄
-showHide=顯示/隱藏標籤
-changeFormatDetail=更改儲存格式
-shapeFillColor=填充顏色
-closeDetail=離開本程式
-copySelectedShape=複製區塊
-delBoxDetail=刪除區塊
-zoomIn=放大畫面
-showInfo=資訊
-showInfoDetail=資訊
-openAnnotation=開啟標籤
-prevImgDetail=上一個圖像
-setFitWidth=縮放到跟畫面一樣寬
-zoomOut=縮小畫面
-changeSaveDirDetail=更改預設標籤存的目錄
-nextImgDetail=下一個圖像
-zoomOrgDetail=放大到原始大小
-prevImg=上一個圖像
-showTutorialDialogDetail=顯示示範內容
-shapeLineColor=形狀線條顏色
-boxLineColor=日期分隔線顏色
-editLabelDetail=修改所選區塊的標籤
-nextImg=下一張圖片
-useDefaultLabel=使用預設標籤
-useDifficult=有難度的
-boxLabelText=區塊的標籤
+saveAsDetail=將標籤保存到其他文件
+closeFile=關
+closeFileDetail=關閉當前文件
+resetAll=重置所有
+resetAllDetail=重置所有
+boxLineColor=箱線顏色
+boxLineColorDetail=選擇框線顏色
+createShape=創建盒子
+createShapeDetail=畫一個新盒子
+setCreateMode=創建盒子
+setCreateModeDetail=畫一個新盒子
+delBox=刪除框
+delBoxDetail=取下盒子
+copySelectedShape=重複盒
+copySelectedShapeDetail=創建所選框的副本
+showTutorialDialog=講解
+showTutorialDialogDetail=顯示演示
+showInfo=信息
+showInfoDetail=應用程式資訊
+zoomIn=放大
+zoomInDetail=增加縮放等級
+zoomOut=縮小
+zoomOutDetail=降低縮放等級
+zoomOrg=原始尺寸
+zoomOrgDetail=縮放到原始大小
+setFitWin=適合窗口
+setFitWinDetail=縮放跟隨窗口大小
+setFitWidth=適合寬度
+setFitWidthDetail=縮放跟隨窗口寬度
+editLabel=編輯標籤
+editLabelDetail=修改所選框的標籤
+shapeLineColor=形狀線顏色
+shapeLineColorDetail=更改此特定形狀的線條顏色
+shapeFillColor=形狀填充顏色
+shapeFillColorDetail=更改此特定形狀的填充顏色
+showHide=顯示/隱藏標籤面板
+useDefaultLabel=使用默認標籤
+useDifficult=難
+boxLabelText=盒子標籤
 labels=標籤
-autoSaveMode=自動儲存模式
-singleClsMode=單一類別模式
-displayLabel=顯示類別
+autoSaveMode=自動保存模式
+singleClsMode=單班模式
+displayLabel=顯示標籤
 fileList=檔案清單
 files=檔案
-advancedMode=進階模式
-advancedModeDetail=切到進階模式
-showAllBoxDetail=顯示所有區塊
-hideAllBoxDetail=隱藏所有區塊
-changeFormat=&PascalVOC
-setEditMode=
-setEditModeDetail=
-commitAnnotatedFrames=提交標籤
-commitAnnotatedFramesDetail=將帶註釋的圖像提交到數據文件夾
+advancedMode=高級模式
+advancedModeDetail=切換至進階模式
+showAllDetail=顯示所有邊界框
+showAll=顯示邊界框
+hideAllDetail=隱藏所有邊界框
+hideAll=隱藏邊界框
+changeFormat=＆PascalVOC
+setEditMode=編輯框
+setEditModeDetail=移動和編輯框
