+++ tests/test_backend.py
@@ -4,7 +4,7 @@ from libs.net.build import TFNet
 from libs.utils.errors import GradientNaN, VariableIsNone
 from libs.dark.darknet import Darknet
 from libs.net.framework import create_framework
-from libs.cython_utils.nms import box_iou
+from libs.cythonUtils.nms import box_iou
 import tensorflow as tf
 
 
@@ -29,7 +29,7 @@ class TestBackend(TestCase, FlagIO):
 
     def testCythonExtensions(self):
         iou = box_iou(3, 4, 4, 3, 4, 3, 3, 4)
-        self.assertAlmostEqual(0.35211268067359924, iou), "cython_utils math failure"
+        self.assertAlmostEqual(0.35211268067359924, iou), "cythonUtils math failure"
 
     def tearDown(self):
         self.cleanup_ramdisk()
